:recycle: Some cleanup
