--- src/frontend/modules/filter/src/components/filters/index.tsx
@@ -1,42 +1,0 @@
-import type { OCELFilter } from "@ocelescope/api-base";
-import type { Control } from "react-hook-form";
-import type { FilterType } from "../../types/filter";
-import {
-  EventAttributeFilter,
-  ObjectAttributeFilter,
-} from "./FilterComponents/AttributeFilter";
-import {
-  EventTypeFilterInput,
-  ObjectTypeFilterInput,
-} from "./FilterComponents/EntityTypeFilter";
-import {
-  E2OCountFilter,
-  O2OCountFilter,
-} from "./FilterComponents/RelationFilter";
-import TimeFrameFilter from "./FilterComponents/TimeFrameFilter";
-
-export type FilterPageComponentProps = {
-  control: Control<OCELFilter>;
-  ocelId: string;
-};
-
-export const filterMap: {
-  [K in FilterType]: {
-    label: string;
-    filterPage: React.ComponentType<FilterPageComponentProps>;
-  };
-} = {
-  event_type: { label: "Event Type", filterPage: EventTypeFilterInput },
-  object_types: { label: "Object Types", filterPage: ObjectTypeFilterInput },
-  time_range: { label: "Time Range", filterPage: TimeFrameFilter },
-  event_attributes: {
-    label: "Event Attributes",
-    filterPage: EventAttributeFilter,
-  },
-  object_attributes: {
-    label: "Object Attributes",
-    filterPage: ObjectAttributeFilter,
-  },
-  e2o_count: { label: "E2O Count", filterPage: E2OCountFilter },
-  o2o_count: { label: "O2O Count", filterPage: O2OCountFilter },
-} as const;
