+++ src/frontend/modules/filter/src/components/inputs/BarChartSelect.tsx
