+++ src/frontend/modules/filter/src/index.ts
@@ -1,3 +1,13 @@
-import ocelotModule from "./config";
+import { defineModule } from "@ocelescope/core";
 
-export default ocelotModule;
+import { FilterIcon } from "lucide-react";
+import FilterPage from "./components/FilterPage";
+
+export default defineModule({
+  name: "filter",
+  description: "Filter functionality provided by the ocelescope package",
+  label: "Filters",
+  authors: [{ name: "Öztürk, Görkem-Emre" }],
+  routes: [FilterPage],
+  icon: FilterIcon,
+});
